move interfaces into a separate directory
